intel-iommu: Default to non-coherent for domains unattached to iommus
Patchwork-id: 42335
O-Subject: [RHEL6.2 PATCH v2] intel-iommu: Default to non-coherent for domains
	unattached to iommus
Bugzilla: 746484
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Chris Wright <chrisw@redhat.com>

V2: (1) Fix $Subject for proper RHEL patch stream i.d.
    (2) Fix bz number
    (3) correct duplication in patch

BZ 746484: Device-assignment broken on SandyBridge EP/EN systems

Backport of following patch posted upstream:

intel-iommu: Default to non-coherent for domains unattached to iommus

domain_update_iommu_coherency() currently defaults to setting domains
as coherent when the domain is not attached to any iommus.  This
allows for a window in domain_context_mapping_one() where such a
domain can update context entries non-coherently, and only after
update the domain capability to clear iommu_coherency.

This can be seen using KVM device assignment on VT-d systems that
do not support coherency in the ecap register.  When a device is
added to a guest, a domain is created (iommu_coherency = 0), the
device is attached, and ranges are mapped.  If we then hot unplug
the device, the coherency is updated and set to the default (1)
since no iommus are attached to the domain.  A subsequent attach
of a device makes use of the same dmar domain (now marked coherent)
updates context entries with coherency enabled, and only disables
coherency as the last step in the process.

To fix this, switch domain_update_iommu_coherency() to use the
safer, non-coherent default for domains not attached to iommus.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Tested-by: Donald Dutile <ddutile@redhat.com>
Acked-by: Donald Dutile <ddutile@redhat.com>
Acked-by: Chris Wright <chrisw@sous-sol.org>

Brew build:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3817473

Testing:
See comment #51 in the bz.
Explains why it wasn't seen in our virt-qe testing:
 > need to modify QE test to assign VF mac-addr in host,
   and match macaddr in guest ifcfg-ethX file so device configures
   into network stack (full dhcp).

Note: Upstream is discussing a much larger change
      in the thread, but the original posting resolves this
      bug with the least amount of risk to RHEL6.

Please review and ack (for final-rc).

- Don

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>