xenfv: mask smep, even if the HV doesn't
Patchwork-id: 42390
O-Subject: [RHEL6.3 PATCH] xenfv: mask smep, even if the HV doesn't
Bugzilla: 752382
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

Bugzilla: 752382
Brew: 3832099
Upstream: N/A

RHEL5 Xen doesn't support smep. For 5.8 we mask the SMEP cpu feature from
the guest. However in order to ensure RHEL6 HVM guests can boot on 5.7
and earlier released hosts we need to mask it in the guest as well. Users
could just boot with nosmep on the command line in these situations, but
we like to reduce the pain the user feels. imammado created a nice cpu
feature filtering function for xenfv guests that we can easily add SMEP
to. Unfortunately the simple addition wasn't enough for this patch, as we
also needed to move the setup of smep down below the filtering. Strangely,
the upstream enable smep commit (de5397ad5b9ad) message says "moved the
call to setup_smep() until after the vendor-specific initialization",
which would be about where I moved it, but the patch doesn't match that,
and the RHEL backport was faithful. So there's a chance upstream would
accept the simple setup_smep move patch, but it would be pretty low
priority. As for the SMEP filtering, that's definitely RHEL specific.

Tested on a machine that supports SMEP with a 5.7 Xen host. The careful
tester will see that we currently only warn N-1 times, where N is the
number of vcpus, that we're disabling SMEP. This of course should be N.
A patch is already posted[1], but not yet integrated, that fixes this.

[1]
Date: Mon, 31 Oct 2011 22:37:30 -0700
From: Chris Wright <chrisw@redhat.com>
To: rhkernel-list@redhat.com, rhvirt-patches@redhat.com
Subject: [RHEL6.2 PATCH 1/2] [x86] cpu: fix CPUID leaf 7 detection
Message-ID: <20111101053730.GP9394@x200.localdomain>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>