nfs: when attempting to open a directory, fall back on normal lookup
Patchwork-id: 42396
O-Subject: [RHEL6.3 PATCH] BZ#755380: nfs: when attempting to open a directory,
	fall back on normal lookup (try #5)
Bugzilla: 755380
RH-Acked-by: Sachin Prabhu <sprabhu@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

This patch fixes a regression that crept in upstream between RHEL5 and
RHEL6. The upstream patch is not in Linus' tree yet, but Trond has
assured me that he's planning to push it to him for 3.2.

Upstream commit: 1788ea6e3b2a58cf4fb00206e362d9caff8d86a7
------------------------------[snip]----------------------------
commit d953126 changed how nfs_atomic_lookup handles an -EISDIR return
from an OPEN call. Prior to that patch, that caused the client to fall
back to doing a normal lookup. When that patch went in, the code began
returning that error to userspace. The d_revalidate codepath however
never had the corresponding change, so it was still possible to end up
with a NULL ctx->state pointer after that.

That patch caused a regression. When we attempt to open a directory that
does not have a cached dentry, that open now errors out with EISDIR. If
you attempt the same open with a cached dentry, it will succeed.

Fix this by reverting the change in nfs_atomic_lookup and allowing
attempts to open directories to fall back to a normal lookup

Also, add a NFSv4-specific f_ops->open routine that just returns
-ENOTDIR. This should never be called if things are working properly,
but if it ever is, then the dprintk may help in debugging.

To facilitate this, a new file_operations field is also added to the
nfs_rpc_ops struct.

Cc: stable@kernel.org
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>