Ivy Bridge kernel rdrand support
Patchwork-id: 42405
O-Subject: [PATCH V2 RHEL6.3] bz696442 Ivy Bridge kernel rdrand support
Bugzilla: 696442
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

This is a trivial backport of the three upstream patches that enabled
use of the rdrand instruction on Ivy Bridge processors.

Brew build:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3801535

Locally tested by me by creating a dummy hwrng module that uses
get_random_bytes() to produce output.  On my Ivy Bridge system, with
this patch:

[root@fenlason-lab3 hw_random]# dd </dev/hwrng bs=4096 count=100>/dev/null
100+0 records in
100+0 records out
409600 bytes (410 kB) copied, 0.114114 s, 3.6 MB/s

With this patch and nordrand on the kernel boot line:
[root@fenlason-lab3 hw_random]# dd </dev/hwrng bs=4096 count=100>/dev/null
100+0 records in
100+0 records out
409600 bytes (410 kB) copied, 0.943388 s, 434 kB/s

So it seems to work.  I have not checked the output for cryptographic
strength.
v2: Now with an important 3-character change.

Upstream:

http://git.kernel.org/?p=linux/kernel/git/torvalds/linux.git;a=commitdiff;h=63d77173266c1791f1553e9e8ccea65dc87c4485
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux.git;a=commitdiff;h=8e6d539e0fd0c2124a20a207da70f2af7a9ae52c
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux.git;a=commitdiff;h=bd29e568a4cb6465f6e5ec7c1c1f3ae7d99cbec1

Locally tested with a fancier dummy hwrng module.  I have not done any
cryptographic tests for randomness, but the results change and are not
all zeros.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>