perf: Fix parsing of __print_flags() in TP_printk()
Patchwork-id: 42418
O-Subject: [PATCH RHEL6.3 BZ730220 2/2] perf: Fix parsing of __print_flags() in
	TP_printk()
Bugzilla: 756311
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 730220
https://bugzilla.redhat.com/show_bug.cgi?id=730220

Description:
============
A update is made to the sched:sched_switch event that adds some
logic to the first parameter of the __print_flags() that shows the
state of tasks. This change cause perf to fail parsing the flags.

A simple fix is needed to have the parser be able to process ops
within the argument.

Upstream status:
================
perf: Fix parsing of __print_flags() in TP_printk()
commit 49908a1b25d448d68fd26faca260e1850201575f
Author: Steven Rostedt <srostedt@redhat.com>
Date:   Fri Nov 4 16:32:25 2011 -0400

Brew:
=====
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3847867

Tested:
=======
tested with instructions in BZ

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>