RAID: Print correct status characters
Patchwork-id: 42521
O-Subject: [RHEL6.3 PATCH] DM RAID: Print correct status characters
Bugzilla: 749684
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Bug #749684:
 https://bugzilla.redhat.com/show_bug.cgi?id=749684

Brew Build:
 http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3855371

Patch Description:
 When devices in a RAID array are not in-sync, they are supposed to be
reported as such in the status output as an 'a' character, which means
"alive, but not in-sync".  But when the entire array is rebuilt 'A' is
being used, which is incorrect.  This patch corrects this to 'a'.

Testing:
Correct operation verified.
[root@bp-01 LVM2]# dmsetup status  --  initial sync (notice 'aa')
vg-lv: 0 2097152 raid raid1 2 aa 1749248/2097152
[root@bp-01 LVM2]# dmsetup status vg-lv  -- after initial sync ('AA')
0 2097152 raid raid1 2 AA 2097152/2097152
[root@bp-01 LVM2]# dmsetup status vg-lv -- after rebuild of one device
0 2097152 raid raid1 2 aA 291200/2097152

Upstream status:
 git ID: 2e727c3ca1beff05f27b6207a795790f222bf8d8

Signed-off-by: Jonathan Brassow <jbrassow@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>