GFS2: Cache dir hash table in a contiguous buffer
Patchwork-id: 42540
O-Subject: [RHEL6.3 GFS2] [Patch 1/2] bz#681906 - GFS2: Cache dir hash table in
	a contiguous buffer
Bugzilla: 681906
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

Hi,

This is the first of two patches back-ported from upstream for
RHEL6.3 bug #681906. One of our business partners has tried the
patch in a test kernel and reported positive results. I tested
it in-house as well on one of our systems.
These are performance improvements to the GFS2 file system.

The upstream patch is here:

http://git.kernel.org/?p=linux/kernel/git/steve/gfs2-3.0-nmw.git;a=commitdiff;h=17d539f0499fa2c0321b7c260831cca2bb36d119

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
--
GFS2: Cache dir hash table in a contiguous buffer

This patch adds a cache for the hash table to the directory code
in order to help simplify the way in which the hash table is
accessed. This is intended to be a first step towards introducing
some performance improvements in the directory code.

There are two follow ups that I'm hoping to see fairly shortly. One
is to simplify the hash table reading code now that we always read the
complete hash table, whether we want one entry or all of them. The
other is to introduce readahead on the heads of the hash chains
which are referred to from the table.

The hash table is a maximum of 128k in size, so it is not worth trying
to read it in small chunks.

rhbz#681906
--

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>