tcp: tcp_syn_retries sysctl is not honored
Patchwork-id: 42587
O-Subject: [RHEL6.3 PATCH] tcp: tcp_syn_retries sysctl is not honored
Bugzilla: 692024
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=692024

This is the same patch as already in RHEL5.7 (bz#688989). However, in RHEL6
the retransmits_timed_out() function resides in include/net/tcp.h. As it is
inline and nothing else uses it, there should be no problem in adding the
necessary parameter.

Note the fix has a minor potential for uncovering hidden bugs in existing
applications as the TCP connection timeout increases (tcp_syn_retries is
5 by default).

Upstream:
commit 4d22f7d372f5769c6c0149e427ed6353e2dcfe61
Author: Damian Lukowski <damian@tvk.rwth-aachen.de>
Date:   Tue, 28 Sep 2010 13:08:32 -0700

net-2.6: SYN retransmits: Add new parameter to retransmits_timed_out()

Fixes kernel Bugzilla Bug 18952

This patch adds a syn_set parameter to the retransmits_timed_out()
routine and updates its callers. If not set, TCP_RTO_MIN is taken
as the calculation basis as before. If set, TCP_TIMEOUT_INIT is
used instead, so that sysctl_syn_retries represents the actual
amount of SYN retransmissions in case no SYNACKs are received when
establishing a new connection.

Brew:
https://brewweb.devel.redhat.com/taskinfo?taskID=3878594

Testing:
By iptables ... -j DROP on a TCP port and netcat to that port.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>