cpu hotplug vs missing pfault completion interrupts
Patchwork-id: 42641
O-Subject: [RHEL6.3 PATCH 01/10] [s390x] kernel: cpu hotplug vs missing pfault
	completion interrupts
Bugzilla: 753080

Description
-----------
Unkillable tasks in an overcommited z/VM environment where cpu
hotplug events happened.

On cpu hot remove a PFAULT CANCEL command is sent to the hypervisor
which in turn will cancel all outstanding pfault requests that have
been issued on that cpu (the same happens with a SIGP cpu reset).
The result is that we end up with uninterruptible processes where
the interrupt that would wake up these processes never arrives.

In order to solve this all processes which wait for a pfault
completion interrupt get woken up after a cpu hot remove. The worst
case that could happen is that they fault again and in turn need to
wait again.

This commit also partially includes the upstream commits
  de5d9bf: Move list types from <linux/list.h> to <linux/types.h>
  88d89da: perf: Add back list_head data types

Bugzilla
--------
BZ 753080
https://bugzilla.redhat.com/show_bug.cgi?id=753080

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 3.0
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=f2db2e6cb3f5f766cbb3788af44705685ff2445a

kABI
----
This patches introduces a new member in the arch-specific thread_info
structure.  This should not matter because this structure is used in
the kernel only.  Also the new member is within a structure part which
is already excluded from the kABI.

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>