cio: disallow driver io for known to be broken paths
Patchwork-id: 42643
O-Subject: [RHEL6.3 PATCH 04/10] [s390x]cio: disallow driver io for known to be
	broken paths
Bugzilla: 765810
RH-Acked-by: Steve Best <sbest@redhat.com>

Description
-----------
Drivers trying to do I/O solely on paths that are online,
but some way defective, might lack the information to do
proper error handling.

There is no reason to allow the usage of known to be
broken paths. Thus restrict the paths a ccw driver can
use for I/O to a subset of the paths cio found usable.

Bugzilla
--------
BZ 765810
https://bugzilla.redhat.com/show_bug.cgi?id=765810

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 3.2
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=659213b899f389d3a1e211796713ceaaa280fa44

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>