ap: Setup timer for sending messages after reset
Patchwork-id: 42642
O-Subject: [RHEL6.3 PATCH 05/10] [s390x]ap: Setup timer for sending messages
	after reset
Bugzilla: 760544
RH-Acked-by: Steve Best <sbest@redhat.com>

Description
-----------
AP messages, which are sent to a busy AP bus device just before a
device timeout, will not be processed after the device recovery.
The pending requests on an AP bus device are not resent to the
device after the device reset.  Use and set up a timer to process
messages in the request queue after a successful AP bus device reset.

Bugzilla
--------
BZ 760544
https://bugzilla.redhat.com/show_bug.cgi?id=760544

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 3.2
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=75464960fc0ccc505527edc1459c8ad191fbc0cc

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>