qeth: improve recovery during resource shortage
Patchwork-id: 42648
O-Subject: [RHEL6.3 PATCH 10/10] [s390x] qeth: improve recovery during resource
	shortage
Bugzilla: 766785
RH-Acked-by: Steve Best <sbest@redhat.com>

Description
-----------
The qeth recovery is not able to fix up the device.  After that,
a recovery is not possible anymore.
If there is a system resource shortage during recovery,
the driver can not start a recovery thread but does not
clear recovery masks to allow recovery again.

In case there are no system resources to run a recovery
clear recovery bitmasks so a further automatic or manual
driven recovery can fix up the device.

Bugzilla
--------
BZ 766785
https://bugzilla.redhat.com/show_bug.cgi?id=766785

Upstream status of the patch
----------------------------
The patch is queued for upstream submission.

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>