crashk_res init check for /sys/kernel/kexec_crash_size
Patchwork-id: 42658
O-Subject: [RHEL6.3 PATCH 07/17] [kdump] crashk_res init check for
	/sys/kernel/kexec_crash_size
Bugzilla: 738862
RH-Acked-by: Steve Best <sbest@redhat.com>

Currently it is possible to set the crash_size via the sysfs
/sys/kernel/kexec_crash_size even if no crash kernel memory has
been defined with the "crashkernel" parameter. In this case
"crashk_res" is not initialized and crashk_res.start = crashk_res.end = 0.
Unfortunately resource_size(&crashk_res) returns 1 in this case.
This breaks the s390 implementation of crash_(un)map_reserved_pages().

To fix the problem the correct "old_size" is now calculated in
crash_shrink_memory(). "old_size is set to "0" if crashk_res is
not initialized. With this change crash_shrink_memory() will do nothing,
when "crashk_res" is not initialized. It will return "0" for
"echo 0 > /sys/kernel/kexec_crash_size" and -EINVAL for
"echo [not zero] > /sys/kernel/kexec_crash_size".

In addition to that this patch also simplifies the "ret = -EINVAL"
or "ret = 0" logic as suggested by Simon Horman.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>