ext4: fix trim length underflow with small trim length
Patchwork-id: 42671
O-Subject: [RHEL6.3 PATCH 05/15] ext4: fix trim length underflow with small trim
	length
Bugzilla: 767219
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

From: Tao Ma <boyu.mt@taobao.com>

BZ 767219
https://bugzilla.redhat.com/show_bug.cgi?id=767219

BREW 3890374
https://brewweb.devel.redhat.com/taskinfo?taskID=3890374

Upstream commit
22f10457432387615fa1ae6e0375d9cacc50819b

In 0f0a25b, we adjust 'len' with s_first_data_block - start, but
it could underflow in case blocksize=1K, fstrim_range.len=512 and
fstrim_range.start = 0. In this case, when we run the code:
len -= first_data_blk - start; len will be underflow to -1ULL.
In the end, although we are safe that last_group check later will limit
the trim to the whole volume, but that isn't what the user really want.

So this patch fix it. It also adds the check for 'start' like ext3 so that
we can break immediately if the start is invalid.

Cc: Lukas Czerner <lczerner@redhat.com>
Signed-off-by: Tao Ma <boyu.mt@taobao.com>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>