ext4: check if device support discard in FITRIM ioctl
Patchwork-id: 42679
O-Subject: [RHEL6.3 PATCH 11/15] ext4: check if device support discard in FITRIM
	ioctl
Bugzilla: 767219
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

BZ 767219
https://bugzilla.redhat.com/show_bug.cgi?id=767219

BREW 3890374
https://brewweb.devel.redhat.com/taskinfo?taskID=3890374

Upstream commit
4143179218960a70d821a425e3c23ce44aa93dee

For a device that does not support discard, the FITRIM ioctl returns
-EOPNOTSUPP when blkdev_issue_discard() returns this error code, which
is how the user is informed that the device does not support discard.

If there are no suitable free extents to be trimmed, then FITRIM will
return success even though the device does not support discard, which
could confuse the user.  So check explicitly if the device supports
discard and return an error code at the beginning of the FITRIM ioctl
processing.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>