blkdev: Submit discard bio in batches in blkdev_issue_discard()
Patchwork-id: 42680
O-Subject: [RHEL6.3 PATCH 14/15] blkdev: Submit discard bio in batches in
	blkdev_issue_discard()
Bugzilla: 767219
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

BZ 767219
https://bugzilla.redhat.com/show_bug.cgi?id=767219

BREW 3890374
https://brewweb.devel.redhat.com/taskinfo?taskID=3890374

Upstream commit
5dba3089ed03f84b84c6c739df8330112f04a15d

Currently we are waiting for every submitted REQ_DISCARD bio separately,
but it can have unwanted consequences of repeatedly flushing the queue,
so we rather submit bios in batches and wait for the entire batch, hence
narrowing the window of other ios going in.

Use bio_batch_end_io() and struct bio_batch for that purpose, the same
is used by blkdev_issue_zeroout(). Also change bio_batch_end_io() so we
always set !BIO_UPTODATE in the case of error and remove the check for
bb, since we are the only user of this function and we always set this.

Remove bio_get()/bio_put() from the blkdev_issue_discard() since
bio_alloc() and bio_batch_end_io() is doing the same thing, hence it is
not needed anymore.

I have done simple dd testing with surprising results. The script I have
used is:

for i in $(seq 10); do
        echo $i
        dd if=/dev/sdb1 of=/dev/sdc1 bs=4k &
        sleep 5
done
/usr/bin/time -f %e ./blkdiscard /dev/sdc1

Running time of BLKDISCARD on the whole device:
with patch              without patch
0.95                    15.58

So we can see that in this artificial test the kernel with the patch
applied is approx 16x faster in discarding the device.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
CC: Dmitry Monakhov <dmonakhov@openvz.org>
CC: Jens Axboe <jaxboe@fusionio.com>
CC: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Jens Axboe <jaxboe@fusionio.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>