blkdev: Do not return -EOPNOTSUPP if discard is supported
Patchwork-id: 42681
O-Subject: [RHEL6.3 PATCH 15/15] blkdev: Do not return -EOPNOTSUPP if discard is
	supported
Bugzilla: 767219
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

BZ 767219
https://bugzilla.redhat.com/show_bug.cgi?id=767219

BREW 3890374
https://brewweb.devel.redhat.com/taskinfo?taskID=3890374

Upstream commit
8af1954d172a46a63e5e79dae523a6d74715e458

Currently we return -EOPNOTSUPP in blkdev_issue_discard() if any of the
bio fails due to underlying device not supporting discard request.
However, if the device is for example dm device composed of devices
which some of them support discard and some of them does not, it is ok
for some bios to fail with EOPNOTSUPP, but it does not mean that discard
is not supported at all.

This commit removes the check for bios failed with EOPNOTSUPP and change
blkdev_issue_discard() to return operation not supported if and only if
the device does not actually supports it, not just part of the device as
some bios might indicate.

This change also fixes problem with BLKDISCARD ioctl() which now works
correctly on such dm devices.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
CC: Jens Axboe <jaxboe@fusionio.com>
CC: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Jens Axboe <jaxboe@fusionio.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>