hpet: Disable per-cpu hpet timer if ARAT is supported
Patchwork-id: 42807
O-Subject: [RHEL6 PATCH BZ 750201 4/4] x86,
	hpet: Disable per-cpu hpet timer if ARAT is supported
Bugzilla: 750201
RH-Acked-by: Rik van Riel <riel@redhat.com>

Backport of linux-2.6 commit 39fe05e58c5e448601ce46e6b03900d5bf31c4b0

If CPU support always running local APIC timer, per-cpu hpet
timer could be disabled, which is useless and wasteful in such
case. Let's leave the timers to others.

The effect is that we reserve less timers.

PRARIT: The above doesn't really explain why this patch is necssary.  If we
only take the first three patches of the patchset we end up in a situation
where the HPET values have not been initialized, but the HPET is marked as
"functional".  When the delayed TSC initialization thread executes, it sees
that the HPET has been marked "functional" and attempts to access the HPET.
Since the HPET is not completely initialized, the HPET code (well ... the
_clocksource_ code) panics.

The "real" effect of this patch is that we move the HPET initializiation
into the hpet_late_init() section of the code, after all of the HPET
values have been initialized.  When the TSC initialization is done after
this and uses the HPET as a reference clock, the code will no longer panic.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>