Fix sendfile write-side file position
Patchwork-id: 42811
O-Subject: [RHEL 6.3] Fix sendfile write-side file position (bz #770023)
Bugzilla: 770023
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Dave Young <dyoung@redhat.com>

The following patch fixes a bug resulting in sendfile writing to the wrong offset
of the destination file descriptor. This is only occurs when the destination
file is seekable (so sockets as destination files work anyway) and if the
write side file position differs from the read side file position.

The patch has come from upstream:

http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=2cb4b05e7647891b46b91c07c9a60304803d1688

This fixes bug #770023

It has been tested with the reproducer listed in the bug.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>