GFS2: Cache the most recently used resource group in the inode
Patchwork-id: 42837
O-Subject: [PATCH 3/9] [RHEL6.3] GFS2: Cache the most recently used resource
	group in the inode
Bugzilla: 747896
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

This is a back port of the upstream commit:

author	Steven Whitehouse <swhiteho@redhat.com>
	Thu, 1 Sep 2011 12:31:59 +0000 (13:31 +0100)
committer	Steven Whitehouse <swhiteho@redhat.com>
	Fri, 21 Oct 2011 11:39:34 +0000 (12:39 +0100)
commit	54335b1fca27b84baa75b1f45985d98262003837

This means that after the initial allocation for any inode, the
last used resource group is cached in the inode for future use.
This drastically reduces the number of lookups of resource
groups in the common case, and this the contention on that
data structure.

The allocation algorithm is the same as previously, except that we
always check to see if the goal block is within the cached rgrp
first before going to the rbtree to look one up.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>