kdump, ioapic: Reset remote-IRR in clear_IO_APIC
Patchwork-id: 42853
O-Subject: [RHEL6.3 PATCH BZ 754335]: x86, kdump,
	ioapic: Reset remote-IRR in clear_IO_APIC
Bugzilla: 754335
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Backport linux-2.6 commit 1e75b31d638d5242ca8e9771dfdcbd28a5f041df

In the kdump scenario mentioned below, we can have a case where
the device using level triggered interrupt will not generate any
interrupts in the kdump kernel.

1. IO-APIC sends a level triggered interrupt to the CPU's local APIC.

2. Kernel crashed before the CPU services this interrupt, leaving
the remote-IRR in the IO-APIC set.

3. kdump kernel boot sequence does clear_IO_APIC() as part of IO-APIC
initialization. But this fails to reset remote-IRR bit of the
IO-APIC RTE as the remote-IRR bit is read-only.

4. Device using that level triggered entry can't generate any
more interrupts because of the remote-IRR bit.

In clear_IO_APIC_pin(), check if the remote-IRR bit is set and if
so do an explicit attempt to clear it (by doing EOI write on
modern io-apic's and changing trigger mode to edge/level on
older io-apic's). Also before doing the explicit EOI to the
io-apic, ensure that the trigger mode is indeed set to level.
This will enable the explicit EOI to the io-apic to reset the
remote-IRR bit.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>