rds_rdma: don't assume infiniband device is PCI
Patchwork-id: 42864
O-Subject: [RHEL6.3 PATCH] rds_rdma: don't assume infiniband device is PCI
Bugzilla: 740149
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>

From: Jes.Sorensen <Jes.Sorensen@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=740149
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=3920943
Upstream: Posted

commit e700cfd8f4c5a40d4a16d7e5f433973495bd6876
Author: Thadeu Lima de Souza Cascardo <cascardo@linux.vnet.ibm.com>
Date: Wed, 4 Jan 2012 20:03:16 -0200

RDS code assumes that the struct ib_device dma_device member, which is a
pointer, points to a struct device embedded in a struct pci_dev.

This is not the case for ehca, for example, which is a OF driver, and
makes dma_device point to a struct device embedded in a struct
platform_device.

This will make the system crash when rds_rdma is loaded in a system
with ehca, since it will try to access the bus member of a non-existent
struct pci_dev.

The only reason rds_rdma uses the struct pci_dev is to get the NUMA node
the device is attached to. Using dev_to_node for that is much better,
since it won't assume which bus the infiniband is attached to.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@linux.vnet.ibm.com>
Cc: dledford@redhat.com
Cc: Jes.Sorensen@redhat.com
Cc: Venkat Venkatsubra <venkat.x.venkatsubra@oracle.com>
Cc: rds-devel@oss.oracle.com
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>